+++ src/maths/ray/ray.c
@@ -19,7 +19,7 @@ t_ray	ray(t_tuple origin, t_tuple direction)
 	return ((t_ray){origin, direction});
 }
 
-t_tuple	position(t_ray ray, float t)
+t_tuple	position(t_ray ray, double t)
 {
 	return (tp_add(ray.origin, tp_mul(ray.dir, t)));
 }
@@ -29,10 +29,10 @@ t_ray	ray_transform(t_ray r, t_mtx4 m)
 	return ((t_ray){mtx_tup_mul(r.origin, m), mtx_tup_mul(r.dir, m)});
 }
 
-t_ray	ray_for_pixel(t_camera cam, float x, float y)
+t_ray	ray_for_pixel(t_camera cam, double x, double y)
 {
-	float const		wx = cam.half_width - (x + 0.5) * cam.pixel_size;
-	float const		wy = cam.half_height - (y + 0.5) * cam.pixel_size;
+	double const		wx = cam.half_width - (x + 0.5) * cam.pixel_size;
+	double const		wy = cam.half_height - (y + 0.5) * cam.pixel_size;
 	t_tuple const	pixel = mtx_tup_mul(point(wx, wy, -1), cam.inverse);
 	t_tuple const	origin = mtx_tup_mul(point(0, 0, 0), cam.inverse);
