+++ src/maths/transformations/translation.c
@@ -12,7 +12,7 @@
 
 #include "matrix.h"
 
-t_mtx4_ret	translation(float x, float y, float z, t_mtx4 res)
+t_mtx4_ret	translation(double x, double y, double z, t_mtx4 res)
 {
 	mtx4_dup(g_identity_matrix, res);
 	res[0][3] = x;
@@ -21,14 +21,14 @@ t_mtx4_ret	translation(float x, float y, float z, t_mtx4 res)
 	return (res);
 }
 
-t_tuple		tp_translation(float x, float y, float z, t_tuple tp)
+t_tuple		tp_translation(double x, double y, double z, t_tuple tp)
 {
 	t_mtx4	rot;
 
 	return (mtx_tup_mul(tp, translation(x, y, z, rot)));
 }
 
-t_mtx4_ret	mx_translation(float x, float y, float z, t_mtx4 res)
+t_mtx4_ret	mx_translation(double x, double y, double z, t_mtx4 res)
 {
 	t_mtx4	buf;
